# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Multitenant SaaS Appointment Booking API
 * OpenAPI spec version: 0.3.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface AdminBusinessSummary {
  id: number;
  slug: string;
  name: string;
  ownerName: string;
  email: string;
  isActive: boolean;
  subscriptionPlan: string;
  maxServicesAllowed: number;
  createdAt: string;
}

export type AdminUpdateBusinessBodySubscriptionPlan =
  (typeof AdminUpdateBusinessBodySubscriptionPlan)[keyof typeof AdminUpdateBusinessBodySubscriptionPlan];

export const AdminUpdateBusinessBodySubscriptionPlan = {
  free: "free",
  basic: "basic",
  pro: "pro",
} as const;

export interface AdminUpdateBusinessBody {
  isActive?: boolean;
  subscriptionPlan?: AdminUpdateBusinessBodySubscriptionPlan;
  maxServicesAllowed?: number;
  name?: string;
  slug?: string;
  ownerName?: string;
  email?: string;
  password?: string;
}

export interface CreateBusinessBody {
  name: string;
  slug: string;
  ownerName: string;
  email: string;
  password: string;
}

export interface BusinessCreatedResponse {
  id: number;
  slug: string;
  name: string;
  ownerName: string;
  email: string;
  password: string;
  createdAt: string;
}

export interface BusinessLoginBody {
  email: string;
  password: string;
}

/**
 * @nullable
 */
export type BusinessProfileThemeMode =
  | (typeof BusinessProfileThemeMode)[keyof typeof BusinessProfileThemeMode]
  | null;

export const BusinessProfileThemeMode = {
  light: "light",
  dark: "dark",
} as const;

export interface BusinessProfile {
  id: number;
  slug: string;
  name: string;
  ownerName: string;
  email: string;
  bufferMinutes: number;
  notificationEnabled: boolean;
  /** @nullable */
  notificationMessage: string | null;
  /** @nullable */
  primaryColor: string | null;
  /** @nullable */
  fontFamily: string | null;
  /** @nullable */
  logoUrl: string | null;
  /** @nullable */
  bannerUrl: string | null;
  /** @nullable */
  themeMode: BusinessProfileThemeMode;
  /** @nullable */
  whatsappApiKey: string | null;
  /** @nullable */
  whatsappPhoneId: string | null;
  googleCalendarEnabled: boolean;
  stripeEnabled: boolean;
  /** @nullable */
  stripePublicKey: string | null;
  subscriptionPlan: string;
  maxServicesAllowed: number;
  isActive: boolean;
  createdAt: string;
}

export interface BusinessLoginResponse {
  token: string;
  business: BusinessProfile;
}

export interface UpdateBusinessProfileBody {
  name?: string;
  ownerName?: string;
  bufferMinutes?: number;
  notificationEnabled?: boolean;
  /** @nullable */
  notificationMessage?: string | null;
}

export interface UpdateBrandingBody {
  /** @nullable */
  primaryColor?: string | null;
  /** @nullable */
  fontFamily?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  /** @nullable */
  themeMode?: string | null;
}

export interface UpdateIntegrationsBody {
  /** @nullable */
  whatsappApiKey?: string | null;
  /** @nullable */
  whatsappPhoneId?: string | null;
  googleCalendarEnabled?: boolean;
  stripeEnabled?: boolean;
  /** @nullable */
  stripePublicKey?: string | null;
}

export interface PublicBusinessInfo {
  id: number;
  slug: string;
  name: string;
  notificationEnabled: boolean;
  /** @nullable */
  notificationMessage: string | null;
  /** @nullable */
  primaryColor: string | null;
  /** @nullable */
  fontFamily: string | null;
  /** @nullable */
  logoUrl: string | null;
  /** @nullable */
  bannerUrl: string | null;
  /** @nullable */
  themeMode: string | null;
  stripeEnabled: boolean;
  /** @nullable */
  depositAmountAgorot: number | null;
}

export interface Service {
  id: number;
  businessId: number;
  name: string;
  price: number;
  durationMinutes: number;
  /** Buffer time after this service */
  bufferMinutes: number;
  /** @nullable */
  imageUrl: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface CreateServiceBody {
  name: string;
  price: number;
  durationMinutes: number;
  /** Buffer time after this service in minutes */
  bufferMinutes?: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface UpdateServiceBody {
  name?: string;
  price?: number;
  durationMinutes?: number;
  bufferMinutes?: number;
  /** @nullable */
  imageUrl?: string | null;
  isActive?: boolean;
}

export interface WorkingHour {
  id: number;
  businessId: number;
  dayOfWeek: number;
  startTime: string;
  endTime: string;
  isEnabled: boolean;
}

export interface WorkingHourInput {
  dayOfWeek: number;
  startTime: string;
  endTime: string;
  isEnabled: boolean;
}

export interface SetWorkingHoursBody {
  hours: WorkingHourInput[];
}

export interface BreakTime {
  id: number;
  businessId: number;
  dayOfWeek: number;
  startTime: string;
  endTime: string;
  /** @nullable */
  label: string | null;
}

export interface BreakTimeInput {
  dayOfWeek: number;
  startTime: string;
  endTime: string;
  /** @nullable */
  label?: string | null;
}

export interface SetBreakTimesBody {
  breaks: BreakTimeInput[];
}

export interface AppointmentDetail {
  id: number;
  businessId: number;
  serviceId: number;
  serviceName: string;
  durationMinutes: number;
  clientName: string;
  phoneNumber: string;
  appointmentDate: string;
  appointmentTime: string;
  /** @nullable */
  notes: string | null;
  status: string;
  createdAt: string;
}

export interface CreatePublicAppointmentBody {
  serviceId: number;
  clientName: string;
  phoneNumber: string;
  /** YYYY-MM-DD */
  appointmentDate: string;
  /** HH:MM */
  appointmentTime: string;
  notes?: string;
}

export interface AvailabilityResponse {
  date: string;
  slots: string[];
  isFullyBooked: boolean;
}

export interface BusinessStats {
  totalAppointments: number;
  todayCount: number;
  thisWeekCount: number;
  upcomingCount: number;
}

export interface CustomerRecord {
  clientName: string;
  phoneNumber: string;
  totalVisits: number;
  /** Total revenue in agorot */
  totalRevenue: number;
  lastVisitDate: string;
  firstVisitDate: string;
}

export interface WaitlistEntry {
  id: number;
  businessId: number;
  /** @nullable */
  serviceId: number | null;
  /** @nullable */
  serviceName: string | null;
  clientName: string;
  phoneNumber: string;
  /** @nullable */
  preferredDate: string | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}

export interface JoinWaitlistBody {
  serviceId?: number;
  clientName: string;
  phoneNumber: string;
  preferredDate?: string;
  notes?: string;
}

export interface RequestUploadUrlBody {
  filename: string;
  contentType: string;
  size: number;
}

export interface RequestUploadUrlResponse {
  uploadUrl: string;
  objectPath: string;
}

export type SuperAdminListBusinessesParams = {
  adminPassword: string;
};

export type SuperAdminCreateBusinessParams = {
  adminPassword: string;
};

export type SuperAdminUpdateBusinessParams = {
  adminPassword: string;
};

export type SuperAdminDeleteBusinessParams = {
  adminPassword: string;
};

export type GetPublicAvailabilityParams = {
  date: string;
  serviceId: number;
};