# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.3.0
  description: Multitenant SaaS Appointment Booking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/business/login:
    post:
      operationId: businessLogin
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessLoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Business Dashboard ────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    BusinessLoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    BusinessLoginResponse:
      type: object
      properties:
        token:
          type: string
        business:
          $ref: "#/components/schemas/BusinessProfile"
      required: [token, business]
    BusinessProfile:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        ownerName:
          type: string
        email:
          type: string
        bufferMinutes:
          type: integer
        notificationEnabled:
          type: boolean
        notificationMessage:
          type: ["string", "null"]
        primaryColor:
          type: ["string", "null"]
        fontFamily:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        themeMode:
          type: ["string", "null"]
          enum: ["light", "dark", null]
        whatsappApiKey:
          type: ["string", "null"]
        whatsappPhoneId:
          type: ["string", "null"]
        googleCalendarEnabled:
          type: boolean
        stripeEnabled:
          type: boolean
        stripePublicKey:
          type: ["string", "null"]
        subscriptionPlan:
          type: string
        maxServicesAllowed:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - slug
        - name
        - ownerName
        - email
        - bufferMinutes
        - notificationEnabled
        - notificationMessage
        - primaryColor
        - fontFamily
        - logoUrl
        - bannerUrl
        - themeMode
        - whatsappApiKey
        - whatsappPhoneId
        - googleCalendarEnabled
        - stripeEnabled
        - stripePublicKey
        - subscriptionPlan
        - maxServicesAllowed
        - isActive
        - createdAt
...